+++ source/Reloaded.Hooks/Internal/AddressRange.cs
@@ -5,16 +5,16 @@
     /// </summary>
     public struct AddressRange
     {
-        public long StartPointer;
-        public long EndPointer;
+        public nuint StartPointer;
+        public nuint EndPointer;
 
-        public AddressRange(long startPointer, long endPointer)
+        public AddressRange(nuint startPointer, nuint endPointer)
         {
             StartPointer = startPointer;
             EndPointer = endPointer;
         }
 
-        public static AddressRange FromStartAndLength(long start, long length)
+        public static AddressRange FromStartAndLength(nuint start, nuint length)
         {
             return new AddressRange(start, start + length);
         }
@@ -22,7 +22,7 @@
         /// <summary>
         /// Returns true if a specified point is contained in this address range.
         /// </summary>
-        public bool Contains(long point)
+        public bool Contains(nuint point)
         {
             return PointInRange(ref this, point);
         }
@@ -30,7 +30,7 @@
         /// <summary>
         /// Returns true if a number "point", is between min and max of address range.
         /// </summary>
-        private bool PointInRange(ref AddressRange range, long point)
+        private bool PointInRange(ref AddressRange range, nuint point)
         {
             if (point >= range.StartPointer &&
                 point <= range.EndPointer)
